Moved classes into appropriate packages
