+++ src/main/java/entity/SpendingSummary.java
@@ -1,4 +1,4 @@
-package ai_insights;
+package entity;
 
 import java.util.Map;
