+++ src/main/java/use_case/ai_insights/GeminiClient.java
@@ -1,4 +1,4 @@
-package ai_insights;
+package use_case.ai_insights;
 
 import java.net.URI;
 import java.net.http.HttpClient;
